+++ packages/pluggable-widgets-tools/test-config/test-index-no-enzyme.js
@@ -1,0 +1,11 @@
+require("@testing-library/jest-dom");
+const { TextEncoder, TextDecoder } = require("util");
+
+Object.defineProperties(global, {
+    TextEncoder: {
+        value: TextEncoder
+    },
+    TextDecoder: {
+        value: TextDecoder
+    }
+});